Refactor <Info> components
No changes were made but the components which render the dataset info (above the viz panels) were broken out into smaller components and (partially) cleaned up.

In general, the styles were shifted into the `<Info>` component so that the child components can focus on returning content.
